Rename `useDisplayName` -> `useGetAccountName`
