+++ src/components/Activity/ActivityDescriptionGovernance.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { getAddress } from 'viem';
 import { useGetMetadata } from '../../hooks/DAO/proposal/useGetMetadata';
 import useAvatar from '../../hooks/utils/useAvatar';
-import useDisplayName from '../../hooks/utils/useDisplayName';
+import { useGetAccountName } from '../../hooks/utils/useGetAccountName';
 import {
   GovernanceActivity,
   MultisigProposal,
@@ -53,7 +53,7 @@ function ProposalAuthor({ activity }: { activity: FractalProposal }) {
         ? getAddress(multisigProposal.confirmations[0].owner)
         : undefined;
 
-  const { displayName: author } = useDisplayName(proposer);
+  const { displayName: author } = useGetAccountName(proposer);
   const avatarURL = useAvatar(author);
 
   if (!proposer) {
