+++ src/components/ui/cards/DAOInfoCard.tsx
@@ -1,7 +1,7 @@
 import { Box, Center, Flex, Link, Text } from '@chakra-ui/react';
 import { Link as RouterLink } from 'react-router-dom';
 import { DAO_ROUTES } from '../../../constants/routes';
-import useDisplayName from '../../../hooks/utils/useDisplayName';
+import { useGetAccountName } from '../../../hooks/utils/useGetAccountName';
 import { useFractal } from '../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../providers/NetworkConfig/NetworkConfigProvider';
 import { SnapshotButton } from '../badges/Snapshot';
@@ -27,7 +27,7 @@ export function DAOInfoCard() {
   const { addressPrefix } = useNetworkConfig();
 
   // for non Fractal Safes
-  const { displayName } = useDisplayName(node?.daoAddress);
+  const { displayName } = useGetAccountName(node?.daoAddress);
 
   // node hasn't loaded yet
   if (!node || !node.daoAddress) {
