+++ src/components/ui/cards/DAONodeInfoCard.tsx
@@ -1,7 +1,7 @@
 import { Center, Flex, FlexProps, Link, Text, VStack } from '@chakra-ui/react';
 import { Link as RouterLink } from 'react-router-dom';
 import { DAO_ROUTES } from '../../../constants/routes';
-import useDisplayName from '../../../hooks/utils/useDisplayName';
+import { useGetAccountName } from '../../../hooks/utils/useGetAccountName';
 import { useFractal } from '../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../providers/NetworkConfig/NetworkConfigProvider';
 import { FractalGuardContracts, FractalNode, FreezeGuard } from '../../../types';
@@ -35,7 +35,7 @@ export function DAONodeInfoCard({ node, freezeGuard, guardContracts, ...rest }: 
   } = useFractal();
   const { addressPrefix } = useNetworkConfig();
   // for non Fractal Safes
-  const { displayName } = useDisplayName(node?.daoAddress);
+  const { displayName } = useGetAccountName(node?.daoAddress);
 
   // node hasn't loaded yet
   if (!node || !node.daoAddress) {
