+++ src/components/ui/menus/AccountDisplay/AccountMenuButton.tsx
@@ -2,7 +2,7 @@ import { Box, Flex, Icon, Show, Text } from '@chakra-ui/react';
 import { CaretDown } from '@phosphor-icons/react';
 import { useTranslation } from 'react-i18next';
 import useAvatar from '../../../../hooks/utils/useAvatar';
-import useDisplayName from '../../../../hooks/utils/useDisplayName';
+import { useGetAccountName } from '../../../../hooks/utils/useGetAccountName';
 import { useFractal } from '../../../../providers/App/AppProvider';
 import Avatar from '../../page/Header/Avatar';
 
@@ -29,7 +29,7 @@ function WalletMenuButton() {
     readOnly: { user },
   } = useFractal();
   const account = user.address;
-  const { displayName: accountDisplayName } = useDisplayName(account);
+  const { displayName: accountDisplayName } = useGetAccountName(account);
   const avatarURL = useAvatar(accountDisplayName);
 
   if (!account) {
